+++ tests/NoSeekStreamTest.php
@@ -29,4 +29,13 @@ class NoSeekStreamTest extends \PHPUnit_Framework_TestCase
         $wrapped->close();
         $this->assertFalse($wrapped->write('foo'));
     }
+
+    public function testCanAttach()
+    {
+        $s1 = Stream::factory('foo');
+        $s2 = Stream::factory('bar');
+        $wrapped = new NoSeekStream($s1);
+        $wrapped->attach($s2->detach());
+        $this->assertEquals('bar', (string) $wrapped);
+    }
 }